requirements.txt
#backend
fastapi~=0.115.12
uvicorn
python-multipart

#ml
pillow~=11.2.1
numpy~=2.1.0
tensorflow~=2.19.0
gdown

#frontend
gradio~=5.30.0
requests~=2.32.3

